# swagger.yml
openapi: 3.0.0
paths:
...
  /api/auth/login:
    post:
      operationId: AuthProxyController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyLoginDto'
      responses:
        '200':
          description: ''
      summary: Authenticate via the identity service and return tokens
      tags: &ref_0
        - auth-proxy
  /api/auth/refresh-token:
    post:
      operationId: AuthProxyController_refreshToken
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyRefreshTokenDto'
      responses:
        '200':
          description: ''
      summary: Rotate refresh token through the identity service
      tags: *ref_0
  /api/auth/logout:
    post:
      operationId: AuthProxyController_logout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyRefreshTokenDto'
      responses:
        '200':
          description: ''
      summary: Revoke the presented refresh token (sign out this device)
      tags: *ref_0
  /api/auth/logout-all:
    post:
      operationId: AuthProxyController_logoutAll
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyRefreshTokenDto'
      responses:
        '200':
          description: ''
      summary: Revoke all refresh tokens for the user
      tags: *ref_0
  /api/auth/me:
    get:
      operationId: AuthProxyController_me
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: Get current user profile with modules from the identity service
      tags: *ref_0
...
  /api/admissions:
    get:
      operationId: AdmissionsController_list
      parameters: []
      responses:
        '200':
          description: ''
      summary: List admissions with status/ward/patient filters
      tags: &ref_8
        - admissions
    post:
      operationId: AdmissionsController_admit
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdmitPatientDto'
      responses:
        '201':
          description: ''
      summary: Admit a patient (optionally with a bed allocation)
      tags: *ref_8
  /api/admissions/{id}:
    get:
      operationId: AdmissionsController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get admission by id
      tags: *ref_8
    patch:
      operationId: AdmissionsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdmissionDto'
      responses:
        '200':
          description: ''
      summary: Update an active admission
      tags: *ref_8
    delete:
      operationId: AdmissionsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Archive an admission
      tags: *ref_8
  /api/admissions/{id}/transfer:
    post:
      operationId: AdmissionsController_transfer
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferAdmissionDto'
      responses:
        '201':
          description: ''
      summary: Transfer an admission to another ward/bed
      tags: *ref_8
  /api/admissions/{id}/discharge:
    post:
      operationId: AdmissionsController_discharge
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DischargeDto'
      responses:
        '201':
          description: ''
      summary: Discharge an active admission
      tags: *ref_8
  /api/appointments:
    get:
      operationId: AppointmentsController_list
      parameters: []
      responses:
        '200':
          description: ''
      summary: List appointments with pagination, date/status filters
      tags: &ref_9
        - appointments
    post:
      operationId: AppointmentsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentDto'
      responses:
        '201':
          description: ''
      summary: Schedule an appointment
      tags: *ref_9
  /api/appointments/{id}/pdf:
    get:
      operationId: AppointmentsController_print
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Print an appointment slip (PDF)
      tags: *ref_9
  /api/appointments/{id}:
    get:
      operationId: AppointmentsController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Get appointment by id
      tags: *ref_9
    patch:
      operationId: AppointmentsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppointmentDto'
      responses:
        '200':
          description: ''
      summary: Update an appointment
      tags: *ref_9
    delete:
      operationId: AppointmentsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Archive an appointment
      tags: *ref_9
  /api/appointments/{id}/check-in:
    post:
      operationId: AppointmentsController_checkIn
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckInAppointmentDto'
      responses:
        '201':
          description: ''
      summary: Check in appointment and start a visit
      tags: *ref_9
  /api/appointments/{id}/cancel:
    post:
      operationId: AppointmentsController_cancel
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelAppointmentDto'
      responses:
        '201':
          description: ''
      summary: Cancel an appointment
      tags: *ref_9
  /api/appointments/{id}/no-show:
    post:
      operationId: AppointmentsController_noShow
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      summary: Mark appointment as no-show
      tags: *ref_9
  /api/appointments/{id}/complete:
    post:
      operationId: AppointmentsController_complete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      summary: Complete an in-progress appointment
      tags: *ref_9
...
info:
  title: EMR API
  description: >-
    Electronic Medical Record: appointments, visits, encounters, dynamic forms,
    and clinical requests
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ProxyLoginDto:
      type: object
      properties:
        username:
          type: string
          example: admin
          description: Username
        password:
          type: string
          example: password
          description: Password (min 6 characters)
      required:
        - username
        - password
    ProxyRefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
...
    AdmitPatientDto:
      type: object
      properties:
        patientId:
          type: string
          description: Patient MRN
        patientName:
          type: string
          description: Patient display name
        wardId:
          type: string
        bedId:
          type: string
          description: Bed to allocate on admission
        admissionDatetime:
          type: string
          default: '2026-09-17T20:30:30.682Z'
        admissionType:
          type: string
          enum: &ref_22
            - EMERGENCY
            - URGENT
            - ELECTIVE
            - MATERNITY
            - SURGICAL
            - MEDICAL
          default: ELECTIVE
        diagnosis:
          type: string
        referringProviderId:
          type: string
        referringProviderName:
          type: string
        notes:
          type: string
      required:
        - patientId
        - patientName
    UpdateAdmissionDto:
      type: object
      properties:
        wardId:
          type: string
        bedId:
          type: string
        admissionDatetime:
          type: string
        admissionType:
          type: string
          enum: *ref_22
        diagnosis:
          type: string
        referringProviderId:
          type: string
        referringProviderName:
          type: string
        notes:
          type: string
    TransferAdmissionDto:
      type: object
      properties:
        wardId:
          type: string
          description: Destination ward
        bedId:
          type: string
          description: Destination bed
    DischargeDto:
      type: object
      properties:
        dischargeType:
          type: string
          enum:
            - DISCHARGED_HOME
            - DIED
            - REFERRED
            - TRANSFERRED
            - SELF_DISCHARGE
            - AMA
          default: DISCHARGED_HOME
        dischargeSummary:
          type: string
        dischargeDatetime:
          type: string
          default: '2026-09-17T20:30:30.683Z'
    CreateAppointmentDto:
      type: object
      properties:
        patientId:
          type: string
        patientName:
          type: string
        appointmentType:
          type: string
          enum: &ref_23
            - CHECKUP
            - FOLLOW_UP
            - CONSULTATION
            - PROCEDURE
            - EMERGENCY
            - SURGERY
            - OTHER
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        providerId:
          type: string
        providerName:
          type: string
        locationId:
          type: string
        scheduleLocation:
          type: string
        priority:
          type: string
          enum: &ref_24
            - ROUTINE
            - URGENT
            - EMERGENCY
          default: ROUTINE
        reason:
          type: string
        notes:
          type: string
      required:
        - patientId
        - appointmentType
        - date
        - startTime
    CheckInAppointmentDto:
      type: object
      properties:
        locationId:
          type: string
        providerId:
          type: string
        providerName:
          type: string
    CancelAppointmentDto:
      type: object
      properties:
        reason:
          type: string
    UpdateAppointmentDto:
      type: object
      properties:
        patientId:
          type: string
        patientName:
          type: string
        appointmentType:
          type: string
          enum: *ref_23
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        providerId:
          type: string
        providerName:
          type: string
        locationId:
          type: string
        scheduleLocation:
          type: string
        priority:
          type: string
          enum: *ref_24
        status:
          type: string
          enum:
            - SCHEDULED
            - CHECKED_IN
            - IN_PROGRESS
            - COMPLETED
            - CANCELLED
            - NO_SHOW
            - MISSED
        reason:
          type: string
        notes:
          type: string
...